feat:fix some issue.

+++ src/AElf.CrossChainServer.Application.Contracts/TokenPool/AddUserLiquidityInfoIndexInput.cs
@@ -7,5 +7,5 @@ public class AddUserLiquidityInfoIndexInput
     public string ChainId { get; set; }
     public Guid TokenId { get; set; }
     public string Provider { get; set; }
-    public long Liquidity { get; set; }
+    public decimal Liquidity { get; set; }
 }
